requirements.txt
openrazer==0.0.1
Pillow==10.0.0
